% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_count.R
\name{diag_count}
\alias{diag_count}
\title{Model Diagnostics for Count Models}
\usage{
diag_count(model, B = 100)
}
\arguments{
\item{model}{A fitted count model object (from fit_ct)}

\item{B}{Number of bootstrap replications}
}
\value{
A list with performance metrics, bootstrap CI, raw bootstrap draws, and model-based CI
}
\description{
Model Diagnostics for Count Models
}
