% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ct.R
\name{fit_ct}
\alias{fit_ct}
\title{Fit count regression models}
\usage{
fit_ct(formula, data, family = "poisson")
}
\arguments{
\item{formula}{Model formula (e.g., y ~ x1 + x2 or y ~ x1 + (1 | group))}

\item{data}{Data frame containing variables in the formula}

\item{family}{Character string specifying model type:
"poisson", "qpois", "nb", "zip", "zinb",
"glmpoisson", or "glmnb"}
}
\value{
A fitted model object with an added field
\code{$ct_family} indicating the chosen model type
}
\description{
Fits a variety of count regression models using a unified interface,
including Poisson, quasi-Poisson, negative binomial, zero-inflated models,
and mixed-effects versions of poisson and negative binomial.
}
\examples{
fit_ct(y ~ x1 + x2, data = df, family = "poisson")
fit_ct(y ~ x1 + (1 | group), data = df, family = "glmpoisson")

}
