% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ct.R
\name{plot_ct}
\alias{plot_ct}
\title{Count Response Plotting}
\usage{
plot_ct(data, response, group = NA, type = "barplot")
}
\arguments{
\item{data}{A data frame}

\item{response}{A string corresponding to the name of the response variable}

\item{group}{Optional string corresponding to grouping variable (default NA)}

\item{type}{type of plot to select from are "boxplot","violin" and "barplot"}
}
\value{
returns a plot, specifically a ggplot2::ggplot object
}
\description{
Count Response Plotting
}
\details{
For plots since count is non-continuous the bar plot makes the most sense
However, for Box plot and violin plots For count data (which is discrete), the density estimate can be misleading or look odd,
it will be more natural if we have a greater range of different discrete values and odd if we have  few unique values.
For the violin plot especially, the smooth kernel density plot  may suggest that the count is continous when it is not, hence we
added the jitter points for both violin and boxplot
}
\examples{

data(mpg,package = "ggplot2")

# Count boxplot
p1 <- plot_ct(mpg, "cyl", type = "boxplot")
p1

# Count boxplot across group
p2 <- plot_ct(mpg, "cyl", group = "class", type = "boxplot")
p2


# Count violin plot
p3 <- plot_ct(mpg, "cyl", type = "violin")
p3


# Count violin plot across group
p4 <- plot_ct(mpg, "cyl", group = "class", type = "violin")
p4


 # Count bar plot
p5 <- plot_ct(mpg, "cyl", type = "barplot")
p5




# Count bar plot across group
p6 <- plot_ct(mpg, "cyl", group = "class", type = "barplot")
p6


}
