% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_data.R
\name{sum_data}
\alias{sum_data}
\title{Title}
\usage{
sum_data(data, response, predictors = None, group = None)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{group}{}
}
\description{
Title
}
