% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ct.R
\name{summary_ct}
\alias{summary_ct}
\title{Count Response Summaries}
\usage{
summary_ct(data, response, group = NA, na.rm = TRUE, plot = FALSE)
}
\arguments{
\item{data}{A data frame}

\item{response}{A string corresponding to the name of the response variable}

\item{group}{group Optional string corresponding to grouping variable (default NA)}
}
\value{
A tibble of summary statistics
}
\description{
Count Response Summaries
}
\details{
Computes basic summary statistics of the count response varaible (5-number summary, variance, variance/ratio,
and zero proportion), optionally grouped by a variable.

Note that count data are technically not continuous summaries need to be interpreted carefully Specifically: The mean of count data can be a non-integer,
even though actual data values are always integers. For variance in count data it's important to compare it to the mean and consider natural dispersion
For median and quantiles it is natural to see repeated or integer jumps in values
}
\examples{

data(mpg, package = "ggplot2")

# Overall summary for the count response
summary_ct(mpg, "cyl")

# Grouped summary for the count response
summary_ct(mpg, "cyl", group = "class")


}
