% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_overall.R
\name{summary_overall}
\alias{summary_overall}
\title{Title}
\usage{
summary_overall(data, na.rm = TRUE, ggpair = TRUE)
}
\arguments{
\item{data}{A data frame}

\item{na.rm}{removing NA when calculating summaries takes on (TRUE/FALSE)}
}
\value{
A tibble of summary statistics
}
\description{
Title
}
\examples{
data(mpg, package = "ggplot2")
summary_overall(mpg)


}
